# service/service/employee/_swagger/service/swagger.yaml
openapi: 3.0.3
info:
  version: "1.0"
  title: "Employee Management"
  contact:
    email: "m.ehab.rabea@gmail.com"
    name: "Mohamed Ehab"
servers:
  - url: http://localhost:7002/
...
paths:
...
  /attendances/reports/daily:
    get:
      tags:
        - EmployeeAttendance
      summary: Get daily attendance report
      operationId: getDailyAttendanceReport
      parameters:
        - $ref: '#/components/parameters/QPAttendanceDate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyAttendanceReport'
        '400':
          $ref: '../../../../../lib/common/_swagger/service/swagger.yaml#/components/responses/BadRequest'

  # ==================== Trainer Course Endpoints ====================
...
components:
  parameters:
...
    QPAttendanceDate:
      name: attendanceDate
      in: query
      required: true
      schema:
        type: string
        format: date
...
  schemas:
...
    EmployeeAttendanceStatus:
      type: object
      properties:
        id:
          type: integer
          enum: [1, 2, 3, 4]
        code:
          type: string
          enum: ["PRESENT", "ABSENT", "LATE", "EXCUSED"]
        arabicLabel:
          type: string
          enum: ["حاضر", "غائب", "متأخر", "معذور"]
...
    EmployeeAttendanceListItem:
      type: object
      properties:
        id:
          type: integer
          format: int32
        employee:
          $ref: '../../../../../service/service/user/_swagger/service/swagger.yaml#/components/schemas/LightUserVTO'
        attendanceDate:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/EmployeeAttendanceStatus'
        checkInTime:
          type: string
          format: time
        checkOutTime:
          type: string
          format: time
        lateTime:
          type: integer
          format: int32
    DailyAttendanceReport:
      type: object
      properties:
        attendanceDate:
          type: string
          format: date
        totalNumber:
          type: integer
          format: int32
        present:
          type: integer
          format: int32
        absent:
          type: integer
          format: int32
        late:
          type: integer
          format: int32
        excused:
          type: integer
          format: int32
        attendanceRate:
          type: number
          format: double
        details:
          type: array
          items:
            $ref: '#/components/schemas/EmployeeAttendanceListItem'
...